updated to user TracerData

+++ src/core/ui/CellBlock.cpp
@@ -3,6 +3,7 @@
 #include <chrono>
 
 #include "imgui.h"
+#include <cmath>
 
 void Tracer::ui::CellBlock::render(bool isFocused, bool isBeingEdited, bool isRowActive) {
     const ImVec2 boxSize = ImVec2(25, 20);
@@ -10,14 +11,14 @@ void Tracer::ui::CellBlock::render(bool isFocused, bool isBeingEdited, bool isRo
     if (isBeingEdited) {
         auto now = std::chrono::steady_clock::now();
         auto time = std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count();
-        float alpha = sinf(time * 0.01f) * 0.1f + 0.7f;
+        float alpha = static_cast<float>(sin(time * 0.01) * 0.1 + 0.7);
         ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.2f, 0.0f, 1.0f, alpha));
         ImGui::Button("--", boxSize);
         ImGui::PopStyleColor();
     } else if (isFocused && !isRowActive) {
         auto now = std::chrono::steady_clock::now();
         auto time = std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count();
-        float alpha = sinf(time * 0.007f) * 0.1f + 0.6f;
+        float alpha = static_cast<float>(sin(time * 0.007) * 0.1 + 0.6);
         ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.0f, 0.0f, 0.0f, 0.0f));
         ImGui::PushStyleColor(ImGuiCol_Border, ImVec4(1.0f, 1.0f, 1.0f, alpha));
         ImGui::PushStyleVar(ImGuiStyleVar_FrameBorderSize, 1.0f);
